requirements.txt
# Core Dependencies
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0.0
PyYAML>=6.0

# AI/LLM
langchain>=0.3
langchain-community>=0.3
langchain-openai>=0.2
langchain-google-genai>=2.0
langgraph>=0.2
langgraph-checkpoint-sqlite>=2.0

# Web
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
requests>=2.31.0
httpx>=0.24.0
beautifulsoup4>=4.12.0
aiohttp>=3.8.0

# CLI/TUI
typer[all]>=0.12.0
rich>=13.0.0
questionary>=2.0.0

# Pentest Tools
sqlparse>=0.4.0

# MCP (Model Context Protocol)
mcp>=1.0.0

# RAG / Database (Optional)
# supabase>=1.0.0
# pgvector>=0.2.0

# Utilities
python-dateutil>=2.8.0

# AI/ML Additional Dependencies
sentence-transformers
supabase
